Make ErrorImpl public and expose it within Error
This change allows differentiating between various types of errors
without resorting to checking the string representation.

Unfortunately, it is not possible to check directly in a `match` on a
`Result` since the `Box` does not automatically dereference in that
situation. You have to do something like:

~~~ rust
match serde_yaml::from_str(&text) {
    Ok(v) => {
        // Handle success
    }
    Err(e) => {
        if let serde_yaml::ErrorImpl::EndOfStream = *(e.0) {
            // Handle EndOfStream error
        } else {
            // Handle all other errors
        }
    }
}
~~~